[feat] Set the ACK to be sender-initiated
And not receiver-initiated like before.
This makes sense because RDMA packets have a bit "ACK requested" in the BTH.
Fix also the logic when ACKs are sent for the last packet, and chunks are ACKed even when the ACK interval is set to zero.
For UD QPs, the logic is not changed and packets are never ACKed.