[fix] Fix for GCC version greather than 6
Fix a bit the source code that prevented a greather version than GCC 6 to compile. The cause is anbiguous `operator<<()`.
To fix it, remove the unused inheritance of `std::ostream`, like in ns-3 recent versions and add overload for `std::vector`.

For some reason, GCC 6 was not complaining but GCC 7 is. And actually it was compiling successfully with GCC 6 (at least in this container), even if said that max. supported version was GCC 5.
For GCC 12, the overload for `std::vector` needed to be added.
Keep GCC 11 because GCC 12 deprecated `std::unary_function` which adds a lot of warnings.

Thanks to that, the container moved to a more recent version of Ubuntu 22.04 and GCC 11. Don't use Ubuntu 24.04 because it removed python2 from apt.