refactor(ns3.19): Mark `override` when possible
This actually changes the behaviour of `QbbChannel::TransmitStart()`, because as a mistake the function has a different `Ptr` template argument (fixed) so it does not override the parent method, but the behaviour seems identical. And actually, there is no point of this class, `PointToPointChannel` does almost he same, the differences being the name of the sink `TxRxQbb` (that is not even used) and the `NS3_MPI` logic in the remote channel.