refactor: Simplify QP keys
SQ QP key is just the source port.
RQ QP key is just the destination port.

Then, the application will need to carefully assign the keys, because this is not done automatically anymore. And the client has now to manually create the RQ QP. This is for the future to permit custom behaviour when sending and receiving packets.